IN PROGRESS: wait_for_motion, detect_motion: Improve visualisation
Instead of highlighting all the part of the screen that is not masked in
red, only the part where motion was actually detected is highlighted.

This allows to visualise what the motion detection algorithm is doing.
However, there is currently duplication between the ` display` part and
`gst_motiondetect_apply`. The two algorithms are equivalent.
Maybe, gst_motiondetect_apply should be written with cvThreshold and
the result image used by display (it will possibly slow down the
detection).
Maybe the old display algorithm should be kept and the display property
could be an integer (0: no display, 1: old visualisation, 2: new
visualisation).