wait_for_all_matches: multiple template matches per frame
This new front end API function lets users specify a list of templates,
all of which must match positively at some point before the `timeout_secs`
expires. This is useful when the object(s) you are attempting to match
appear and disappear from the video source too quickly for 2 sequential
`wait_for_matches`.

Instead, this function will grab a frame, match all unmatched templates
against the frame, and repeat until (1) all templates have been matched or
(2) `timeout_secs` expires. The returned object contains key/value pairs
of the template's filename and the positive MatchResult it returned when
its match was detected.

Users should note that the more templates given to be matched, the less
frames will be matched against per second.

This commit adds a new exception to match the matching function:
`MatchAllTimeout`. This is raised by `wait_for_all_matches` when it fails
to match all the specified templates. However, for ease of debugging, it
lists all the images it *did* match as well as all those it *didn't*.