Fix running `make check` when stbt is not installed
Some of the integration tests would fail if stbt was not installed.  This
means that the tests would have been running against (or at least affected
by) the installed version of stbt.  This is bad as the results from `make
check` are then misleading - the installed version may pass but the source
version which you want to check fails.

This commit fixes this issue by taking the sledgehammer approach - do a
`make install` into a sandboxed directory before running the tests.  This
has the beneficial side effect that the integration tests are now also
testing that `make install` works correctly and issues like this are
now unlikely to recur.

This is required for Debian packaging as Debian runs `make check` while
building a package, but is a good idea anyway.