stbt batch run: Fix stbt-version.log
Since 1dbb8ffa, `stbt batch run` wouldn't correctly record the stbt
version when stbt was run from the installed location (it only worked
when run from source directory). Instead, you'd see this on stderr:

    $prefix/libexec/stbt/stbt-batch.d/../stbt: No such file or directory

We can't find the `stbt` executable relative to `stbt-batch.d/run`,
because the former is in $bindir while the latter is in $libexecdir, so
the location depends on the packaging -- at most we could make an
educated guess.

Note that the selftest I've added does correctly detect the issue,
because since 4ffa6b6c `make check` installs stbt into a temporary
directory and runs the tests against that installation.