stbt batch run: Support test scripts not in a git repository
Test scripts not in a git repository are supported, but the features
offered by `stbt batch` are degraded: The test script's revision will
not appear in the report, and the test name in the report will be the
test script's absolute path (when in a git repo, only the path inside
the repository is shown).

We are not going to explicitly support subversion, mercurial, or any
other version-control system -- see
https://github.com/drothlis/stb-tester/issues/96.
Git does have two-way interoperability with subversion and mercurial, so
if you want these enhanced features of `stbt batch` use git to deploy
your test scripts from your subversion/mercurial repository to your
stb-tester rig.

Closes #96.