test_that_video_index_is_written_on_eos is robust under load
When running tests in parallel, this test would occasionally fail with:

    File "webminspector/webminspector.py", line 116, in <module>
      main()
    File "webminspector/webminspector.py", line 97, in main
      pos = segment.SearchLevelOneElements(file_, pos)
    File "webminspector/webm.py", line 560, in SearchLevelOneElements
      size, start_pos = ReadOneByteForElementIdSize(file_, pos)
    File "webminspector/webm.py", line 53, in ReadOneByteForElementIdSize
      file_.seek(pos)
    TypeError: 'NoneType' object cannot be interpreted as an index

Presumably `stbt run` didn't manage to send sufficient data for the
video to be written properly.

This commit re-runs the test with a longer `sleep` if it fails the first
time. I didn't simply increase the timeout because this test doesn't
usually fail when running the tests sequentially (if you don't have GNU
`parallel` installed), so I didn't want to increase the run-time of
a sequential `make check` unnecessarily.